requirements.txt
flask
flask-wtf
python-dotenv
flask-sqlalchemy
flask-migrate
pymysql
werkzeug
flask-login
ipython

